\name{sklearnPCA-package}
\alias{sklearnPCA-package}
\alias{sklearnPCA}
\docType{package}
\title{
\packageTitle{sklearnPCA}
}
\description{
\packageDescription{sklearnPCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sklearnPCA}
\packageIndices{sklearnPCA}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sklearnPCA}

Maintainer: \packageMaintainer{sklearnPCA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
